requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
openai
faiss-cpu
pydantic
requests
tiktoken
python-dotenv